refactor: 폴더 구조 변경 및 suspense 적용
