+++ src/recoil/selectedCardItems.ts
@@ -1,47 +1,47 @@
-import { atomFamily, selector, selectorFamily } from 'recoil';
-import { recoilPersist } from 'recoil-persist';
-import { CartItemsSelector, CartItemsState } from './cartItems';
+import { atomFamily, selector } from 'recoil';
+import { fetchedCartItemsState } from './cartItems';
+import { localStorageEffect } from '../util/localStorage';
 
-const { persistAtom } = recoilPersist({
-  key: 'cartItem',
-  storage: localStorage,
-});
-
-export const SelectedCartItems = atomFamily<boolean, number>({
+export const selectedCartItemsState = atomFamily<boolean, number>({
   key: 'cartItem',
   default: true,
-  effects_UNSTABLE: [persistAtom],
+  effects_UNSTABLE: (param) => [localStorageEffect(`cartItem_${param}`, true)],
 });
 
-export const SelectedAllCardItemsSelector = selectorFamily<boolean, number[]>({
-  key: 'selectedAllCardItems',
-  get:
-    (cartItemIds) =>
-    ({ get }) => {
-      return cartItemIds.every((itemId) => get(SelectedCartItems(itemId)));
-    },
-  set:
-    (cartItemIds) =>
-    ({ set }, newValue) => {
-      cartItemIds.forEach((itemId) => {
-        set(SelectedCartItems(itemId), newValue);
-      });
-    },
-});
-
-export const SelectedSomeCardItemsSelector = selector<boolean>({
+/**
+ * 카트 상품 전체 선택
+ * get: 각 리스트의 상품들이 전부 선택되었는지를 판단 .
+ * set: 카트 상품 전체 선택 버튼 클릭 시 각 상품들 선택 취소 or 선택
+ */
+export const selectedAllCartItemState = selector<boolean>({
   key: 'selectedAllCardItems',
   get: ({ get }) => {
-    get(CartItemsState);
-    return get(CartItemsId).some((cartItemId) =>
-      get(SelectedCartItems(cartItemId)),
+    return get(cartItemsIdState).every((itemId) =>
+      get(selectedCartItemsState(itemId)),
+    );
+  },
+  set: ({ set, get }, newValue) => {
+    get(cartItemsIdState).forEach((itemId) => {
+      set(selectedCartItemsState(itemId), newValue);
+    });
+  },
+});
+
+/**
+ * 모든 카트 상품들이 선택 해제되었는지 확인
+ */
+export const selectedSomeCartItemsState = selector<boolean>({
+  key: 'selectedSomeCardItems',
+  get: ({ get }) => {
+    return get(cartItemsIdState).some((cartItemId) =>
+      get(selectedCartItemsState(cartItemId)),
     );
   },
 });
 
-export const CartItemsId = selector({
+export const cartItemsIdState = selector({
   key: 'cartItemsId',
   get: ({ get }) => {
-    return get(CartItemsSelector).map((cartItem) => cartItem.id);
+    return get(fetchedCartItemsState).map((cartItem) => cartItem.id);
   },
 });
